pull compare back into reporter, change reporter exports
