+++ bin/index.js
@@ -2,7 +2,7 @@
 
 const { inspect } = require('util')
 const files = require('../src/files')
-const reporter = require('../src/reporter')
+const { reporter } = require('../src/reporter')
 const build = require('../src/build')
 const input = require('../src/input')
 const api = require('../src/api')
